+++ lib/titan.rb
@@ -1,5 +1,6 @@
 require "titan/version"
 
 module Titan
+  autoload :Cli,      "titan/cli"
   autoload :Thread,   "titan/thread"
 end